refactor(dashboard-v2): migrate variables bar to the overhauled model
